refactor: updating user allergens moved into a composable
